prov/efa: Open extended CQ in rxr_ep
In a previous commit, a new member ibv_cq_ex was added to rxr_ep
(https://github.com/ofiwg/libfabric/pull/8035)
But only ibv_create_cq was used to open the CQ

This commit adds support for extended CQ which is created by calling
efadv_create_cq
