+++ prov/efa/src/efa_cq.c
@@ -36,6 +36,7 @@
 #include "config.h"
 #include <ofi_mem.h>
 #include "efa.h"
+#include "efa_cq.h"
 #include <infiniband/verbs.h>
 
 static inline uint64_t efa_cq_opcode_to_fi_flags(enum ibv_wc_opcode	opcode) {
@@ -251,102 +252,24 @@ static struct fi_ops efa_cq_fi_ops = {
 };
 
 /**
- * @brief Create ibv_cq_ex by calling ibv_create_cq_ex
- *
- * @param[in] cq Pointer to the efa_cq.
- * @param[in] attr Pointer to fi_cq_attr.
- * @param[out] Return pointer to ibv_cq_ex if successful, otherwise NULL
- */
-static inline struct ibv_cq_ex *efa_cq_ibv_create_cq_ex(struct efa_cq *cq, struct fi_cq_attr *attr) {
-	struct ibv_cq_init_attr_ex init_attr_ex = {
-		.cqe = attr->size ? attr->size : EFA_DEF_CQ_SIZE,
-		.cq_context = NULL,
-		.channel = NULL,
-		.comp_vector = 0,
-		/* EFA requires these values for wc_flags and comp_mask.
-		 * See `efa_create_cq_ex` in rdma-core.
-		 */
-		.wc_flags = IBV_WC_STANDARD_FLAGS,
-		.comp_mask = 0,
-	};
-
-	return ibv_create_cq_ex(cq->domain->device->ibv_ctx, &init_attr_ex);
-}
-
-/**
  * @brief Create and set cq->ibv_cq_ex
  *
  * @param[in] cq Pointer to the efa_cq. cq->ibv_cq_ex must be NULL.
  * @param[in] attr Pointer to fi_cq_attr.
  * @param[out] Return code = 0 if successful, or negative otherwise.
  */
-#if HAVE_EFADV_CQ_EX
-static inline int efa_cq_set_ibv_cq_ex(struct efa_cq *cq, struct fi_cq_attr *attr) {
-	struct ibv_cq_init_attr_ex init_attr_ex = {
-		.cqe = attr->size ? attr->size : EFA_DEF_CQ_SIZE,
-		.cq_context = NULL,
-		.channel = NULL,
-		.comp_vector = 0,
-		/* EFA requires these values for wc_flags and comp_mask.
-		 * See `efa_create_cq_ex` in rdma-core.
-		 */
-		.wc_flags = IBV_WC_STANDARD_FLAGS,
-		.comp_mask = 0,
-	};
-
-	struct efadv_cq_init_attr efadv_cq_init_attr = {
-		.comp_mask = 0,
-		.wc_flags = EFADV_WC_EX_WITH_SGID,
-	};
+static inline int efa_cq_set_ibv_cq_ex(struct efa_cq *cq, struct fi_cq_attr *attr)
+{
+	enum ibv_cq_ex_type ibv_cq_ex_type;
 
 	if (cq->ibv_cq_ex) {
 		EFA_WARN(FI_LOG_CQ, "CQ already has attached ibv_cq_ex\n");
 		return -FI_EALREADY;
 	}
 
-	/* To enable additional WC fields, create CQ with efadv_create_cq verb instead. */
-	cq->ibv_cq_ex = efadv_create_cq(cq->domain->device->ibv_ctx,
-						   &init_attr_ex,
-						   &efadv_cq_init_attr,
-						   sizeof(efadv_cq_init_attr));
-
-	if (cq->ibv_cq_ex) {
-		/* Support operations on EFA DV CQ */
-		cq->flags |= EFA_CQ_SUPPORT_EFADV_CQ;
-		return 0;
-	}
-
-	/* This could be due to old EFA kernel module versions */
-	EFA_WARN(FI_LOG_CQ, "Unable to create EFA DV CQ: %s\n", strerror(errno));
-	cq->flags &= ~EFA_CQ_SUPPORT_EFADV_CQ;
-
-	/* Fallback to ibv_create_cq_ex */
-	cq->ibv_cq_ex = efa_cq_ibv_create_cq_ex(cq, attr);
-
-	if (!cq->ibv_cq_ex) {
-		EFA_WARN(FI_LOG_CQ, "Unable to create extended CQ: %s\n", strerror(errno));
-		return -FI_ENOCQ;
-	}
-
-	return 0;
+	return efa_cq_ibv_cq_ex_open(attr, cq->domain->device->ibv_ctx,
+				    &cq->ibv_cq_ex, &ibv_cq_ex_type);
 }
-#else
-static inline int efa_cq_set_ibv_cq_ex(struct efa_cq *cq, struct fi_cq_attr *attr) {
-	if (cq->ibv_cq_ex) {
-		EFA_WARN(FI_LOG_CQ, "CQ already has attached ibv_cq_ex\n");
-		return -FI_EALREADY;
-	}
-
-	cq->ibv_cq_ex = efa_cq_ibv_create_cq_ex(cq, attr);
-
-	if (!cq->ibv_cq_ex) {
-		EFA_WARN(FI_LOG_CQ, "Unable to create extended CQ: %s\n", strerror(errno));
-		return -FI_ENOCQ;
-	}
-
-	return 0;
-}
-#endif
 
 int efa_cq_open(struct fid_domain *domain_fid, struct fi_cq_attr *attr,
 		struct fid_cq **cq_fid, void *context)
